@@ -56,31 +59,32 @@
     "native-base": "2.13.8",
     "node-fetch": "^2.6.7",
     "prop-types": "^15.7.2",
-    "react": "16.13.1",
-    "react-dom": "16.13.1",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-42.0.0.tar.gz",
+    "react": "17.0.1",
+    "react-dom": "17.0.1",
+    "react-native": "0.64.3",
     "react-native-appearance": "~0.3.3",
     "react-native-calendars": "^1.264.0",
     "react-native-chart-kit": "^5.5.0",
     "react-native-country-picker-modal": "^2.0.0",
     "react-native-datepicker": "^1.7.2",
-    "react-native-gesture-handler": "~1.10.2",
+    "react-native-easy-toast": "^2.0.0",
+    "react-native-gesture-handler": "~2.1.0",
     "react-native-global-props": "^1.1.5",
     "react-native-hide-with-keyboard": "^1.2.1",
     "react-native-keyboard-aware-scroll-view": "^0.9.3",
     "react-native-keyboard-spacer": "^0.4.1",
-    "react-native-maps": "0.28.0",
+    "react-native-maps": "0.29.4",
     "react-native-modal": "^11.5.6",
     "react-native-modal-datetime-picker": "^8.9.3",
     "react-native-multiple-select": "^0.5.5",
-    "react-native-pager-view": "~5.0.12",
+    "react-native-pager-view": "5.4.9",
     "react-native-paper": "^4.7.2",
     "react-native-popup-menu": "^0.15.10",
+    "react-native-reanimated": "~2.3.1",
     "react-native-raw-bottom-sheet": "^2.2.0",
-    "react-native-reanimated": "~2.2.0",
     "react-native-responsive-screen": "^1.4.1",
-    "react-native-safe-area-context": "3.2.0",
-    "react-native-screens": "~3.4.0",
+    "react-native-safe-area-context": "3.3.2",
+    "react-native-screens": "~3.10.1",
     "react-native-select-input-ios": "^2.0.4",
     "react-native-snap-carousel": "^3.8.4",
     "react-native-svg": "12.1.1",
