@@ -2,8 +2,8 @@
 
 import * as types from '../types'
 
-import { AsyncStorage } from 'react-native'
 import * as SplashScreen from 'expo-splash-screen'
+import AsyncStorage from '@react-native-async-storage/async-storage'
 import axios from 'axios'
 import config from '../config'
 // import * as Sentry from 'sentry-expo'
